Add GitLab-Shell distributed tracing
GitLab Shell is unable to configure itself through environment
variables.

For this reason, we need to configure distributed tracing in
GitLab-Shell's config.yml file.

This means a refactor to remove the hardcoded environment variables,
which were in lib/run.rb as we also needed this value to be written to
`gitlab-shell/config.yml`.

It now uses Foreman environment files. This allows us to manage the
configuration through make without having to change the ruby files. It
also gives the user a bit more flexibility in configuring the values.

On startup, Foreman will read environment values from `gdk.env`. These
are preconfigured with the tracing values, but can also be configured
with other values (for debugging purposes for example)