Update prometheus docs with daily workflow
Minikube can be pretty high-strung, which costs a lot of time and
energy for the folks using it. By documenting a daily workflow
and providing scripts to make setup and usage easier, we hope to
avoid the need to spend time debugging.

This commit provides shell commands for creating, starting,
stopping, listing, backing up, and restoring minikube machines.
It also documents the daily workflow for using minikube.