Add rake gitlab:mr:checkout to checkout a MR
GitLab has multiple components, and these need to be stopped and started
in a correct order when switching to another branch, while also some
steps might have to be taken too. These steps include running yarn, or
bundle.

`rake gitlab:mr:checkout` aims to provide a failure proof way of going
from one merge request on `gitlab-org/gitlab` to another. Internally it
does this by using the `merge-requests/<iid>/head` refspec with a local
mapping. Beside the Git aspect, this command depends on Runit as it
provides great control over the services. For the command reliability is
proffered over speed, which means that lots of tasks are executed which
might or might not have to be executed.