Few improvement to run GDK in the cloud
Use listen address instead of hostname in some places because the hostname
address might be bound to an IP that the machine in the cloud might not have
direct access to. So the cloud computer will set `0.0.0.0` as listen address,
while the hostname is some domain that will resolve to a public IP on the
interwebs.