Correctly test if PostgreSQL is in recovery
When using `ifeq` to define a recipe of rule, these are set during
evaluation of the Makefile. But in fact we should test the state of
PostgreSQL recovery at the moment we want to execute the recipe of
`gitlab-db-migrate`.

So use some shell magic while executing the recipe.

A thorough explanation of the issue is described at:
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/995#note_377316256

Closes https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/995