Use Ruby code to determine which migrations to run
The solution in
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1407
works, but it ain't pretty. It will also hard to maintain.

So instead this introduces the same logic, but nicer written in Ruby
code.

It also will always us to check if the database is in recovery, which is
useful to decide where to put the praefect database. See:
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/984