clone: Replace shallow clone with filter spec
Git clones can take a long time when there's a lot of object to fetch
from a remote server. Usually it's also a lot of time spend on fetching
large blobs from the server. For example all image revisions.

Git has an extension, git-lfs, which partially fixes this problem.
Another option is to use a native feature of Git, partial clones. This
change adds a simple blob filter. Which means that the server will not
send all blobs to client, if they exceed a size the client set. This
works for all Git versions after 2.22.

This means however, that Git has objects missing from the packs. These
are now pretended to be present in so called promisor pack files. When
Git needs a blob when for example running `git checkout <old revision>`
it will fetch it on demand. Given there's no user interaction required,
this is more seemless than shallow clones. Which is why this change
removes that feature.