doc: fix a broken forum link
This link was a 404, maybe it worked at some point, but not anymore.

This seems to have been broken in 773ccb73 ("Apply suggestion to
README.md", 2020-05-27) as a result of overzelous URL
shortening.

Let's not make the same mistake again and use
https://forum.gitlab.com/c/community/community-contributions/15 even
if https://forum.gitlab.com/c/community/community-contributions/ also
works for the moment.