gdk bash completion: make the completion work
Fix the broken bash completion. This was added back in 9644b845 ("Add
Bash completion for the gdk command", 2020-01-23) and worked then (I
assume), but as noted in [1] was broken for zsh in f089a3c7 ("Fix
shellcheck violations", 2020-04-21)[2], and then in 39113cca ("Fix
autocomplete for zsh", 2020-10-02)[3] bash broke when zsh's was fixed.

Fix it to work as intended. The COMPREPLY variable must be filled as
an array, so that:

    $ gdk <TAB>

Produces output like:

    $ gdk
    init	trust	version
    $ gdk <cursor>

Instead, before this, it would produce:

    $ gdk version
    init
    trust <cursor>

I.e. the COMPREPLY that should be an array of options was being set to
a \n-delimited string. The main functional change is small, I just
re-indented the "case" block.

1. https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1611
2. https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1190
3. https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1590