gdk bash completion: support "gdk redis-cli <tab>"
Add completion for the "redis-cli" subcommand by farming it off to the
native "redis-cli" completion on the system. Calling _command_offset
on ZSH would be an error, so let's exempt it.

The previous behavior was for "gdk redis-cli<TAB>" to complete
nothing, which is still the ZSH behavior.

We could support ZSH here by doing whatever _command_offset is doing,
but I couldn't find a quick way to list redis-cli's supported
commands, so I'm leaving that for someone who cares.

Helped-by: https://github.com/99designs/aws-vault/issues/70 &
           https://github.com/segevfiner/aws-vault/commit/4340858d18c9c484d19c911c0de9641fc7b75588