doc: remove duplication from bootstrap & instructions
Add a `make bootstap-packages` target which does just the `apt-get`
part of the bootstrap. The manual installation instructions had to
copy/paste the packages_debian.txt list. Now we're just noting the
packages unique to the asdf-less installation.

This could be improved by having all the package lists be in *.txt,
adding Ubuntu support etc. I'm just doing Debian for now.

This includes a change from Ash McKenzie discussed in [1] to add "ed"
to packages.txt and packages_debian.txt. It's needed because the
Makefile shells out to "ed" for some changes. See [2]. By adding it we
ensure that the "make bootstrap-packages" step will install it.

1. https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1656
2. https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1656#note_443258871