doc: more detailed Debian instructions for the manual installation
The advice to use Go from experimental is outdated, these days the Go
is too new. See https://gitlab.com/gitlab-org/gitaly/-/issues/3036

Update the docs to show how to check for the version you need and
search for it in Debian, do the same for the rbenv instructions. It
doesn't have the currently required 2.7.2 as a version you can select.