pre-flight checks: fix the parse_version() regexp
The regexp wasn't quoting "." so it matched anything instead of the
dot. This broken regex has been here for a while, and seems to have
been copied around a bit, although this is the last occurrence of it:

    git log --reverse -p -G'\\d\+\.\\d\+'

This parsed "Yarn v0.1.1 2011" as "0.1.1 2011" instead of "0.1.1", the
"2011" was the start of a copyright line. Why did I have a 2011
version of Yarn? I had a broken Debian setup, so I doubt this'll
happen in practice with any of our tools, but let's fix the regex
anyway.

Now it'll match numbers split by dots, and not across whitespace or
whatever.