gdk doctor: don't run some diagnostics in parallel
Fixes a regression in 700a92a1 ("Run diagnostics in parallel",
2020-04-15) where we'd intermittently report the need to run
migrations for gitlab/. The cause is that we run the "Configuration",
which in turn runs DiffConfig.execute(). That method racily does the
equivalent of:

    mv cfg cfg.backup &&
    make cfg &&
    diff cfg.backup cfg

Which races with e.g.:

    bundle exec rails db:abort_if_pending_migrations

Which'll fail not because there are outstanding migrations, but
because no database/config.yml exists.

This fix does not try to address the actual race condition, but rather
assumes we'll have hard-to-fix races and will want to deal with them
by just splitting up the parallel from the non-parallel checks. A
subsequent commit will try a more minimal fix by dealing with the race
condition itself.

On my installation this about doubles the "gdk doctor" runtime from
7.2 to 14.4 seconds.