gdk doctor: don't race on config changes
Since 700a92a1 ("Run diagnostics in parallel", 2020-04-15) we've had
all sorts of weird race conditions on checks that need our config
files, with this code that moves them out of the way and re-generates
them.

In particular we'd erroneously report that we had outstanding
migrations because our "do we have them?" command would return
non-zero because the database config file didn't exist, not because we
had outstanding migrations.

One way to fix this would be to simply split up the non-parallel and
parallel checks, but the config step takes so long that "gdk doctor"
with that solution runs in 2x the time. See the commit 2 commits back
for that solution.

This is a more isolated change, but has the negative effect that we
might have trickier races in the futrue so we'll need that
some-serial-some-parallel patch anyway, and worse if checks want to
check the newly generated config values "gdk doctor" might need to run
twice to report the right results.

I think 2x-ing the runtime sucks more than those potential issues, so
let's just copy instead of moving the backup file, and generate the
new one in-place.