Add support for using TCP ports with Rails
By default, the GDK will listen to a UNIX domain socket
(e.g. `/home/git/gdk/gitlab.socket`), but this makes debugging with
tools such as Wireshark difficult.

Now, it's possible to specify a TCP port by:

```yaml
gitlab:
  rails:
    address: gdk.test:3001
```

Note that http:// is omitted because Puma can only listen to TCP
port for HTTP traffic.

This was extracted from
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/1292.

Note this commit also changes gitlab-shell to communicate through
Workhorse, which we do now in Omnibus and the Helm Chart. This
simplifies the gitlab-shell config.

Closes https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/889