Major refactor on ConfigSettings
Prior to this change, the ConfigType classes where used to parse the
initial value inside the yaml. But this makes is impossible to validate
values when they are changed. This would be an issue for:
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/939

With this refactor the ConfigType objects are retained. This puts more
responsibility in these type classes. Now they will be responsible to
parse and dump each individual setting. Later on we can also use them to
dump only the user-configured YAML values and validate the values when
setting them.