Further refactor GDK::ConfigSettings
In the previous refactor there was a difference made between a class
that is initialized during the loading of the class and when reading
values. Both types of the object depended on which DSL method was
called.

This change simplifies things again. There is now one generic Builder
class. Instances of this are made when calling the DSL methods. The
builder holds the type of the object, but the builder objects itself are
generic. This makes the definition of the ConfigTypes again simpler.

With this change the objects know if a user value is set and what the
default value is. This will make it easier later on to dump a YAML
containing only user values.