settings_klass is no longer a class method
In the previous commit we refactored code from the commit leading to
that. That made the change to have a class method settings_klass
superfluous.

This commit rolls back the change to make is a class method, but it
needs to be public, where it could be private before this whole refactor
started.