praefect: Use JSON-formatted logging
When not configuring any logging format in either Gitaly or Praefect,
then we default to a mostly unstructured "text" logging format. This
format is missing fields which would be present in other logging formats
and is thus suboptimal. This is why the GDK already configures Gitaly to
use the JSON log format, but for Praefect it leaves the default logging
format in place.

Let's change Praefect to also use JSON logging to fix this shortcoming
and be consistent across both Gitaly and Praefect.