Minimal Changes Necessary to Use a Nix Flake
Currently only starts `rails-web`, `rails-background-jobs` and `webpack`
via `nix`. More services are dependent on managing their dependencies
via `nix`. Easy next targets are `redis` and `postgres`, pretty much
everything that is a binary that is vendored and does not need to be built.

Also probably MacOS only, would need to do some testing on Linux to
ensure the right dependencies are passed in.