redis: Ignore snapshots dumps
Recently redis started to create a dump.rdb when it is stopped. You can
see this happen when you run `gdk start redis` and `gdk stop
redis`. This dump shouldn't make the GDK repository dirty.

This change adds the file to .gitignore. And because the dump belongs to
redis, the configuration file is changed to save that file in the redis/
subdirectory of GDK.