Fix the path used in make postgres-replication/data
The current make task is creating the data directory which is already in
use by the primary database so it will fail when it already exists.
Based on the docs at `doc/howto/database_load_balancing.md` it seems
like this is meant to generate a directory called
`postgres-replica/data` so we've updated it to match those docs.