Improve gitpod workspace image build
This is a major change to how we build the image, instead of basing our
workspace image on gitpod/workspace-full we are now utilizing the much
smaller gitpod/workspace-base which doesn't ship things like e.g. java.

Furthermore we are now utilizing `asdf` and our `make bootstrap`
installation. This reduces the image size immensely, hopefully leading
to faster startup times. We went down from 5.7 GB => 2.9 GB.

Additionally we are preparing the CI pipeline for scheduled builds so
that we can update the gitpod workspace image more often.