GDK::ConfigType: Imitate autovivification in yaml
To make it possible to write user-defined values to the yaml in sections
that not exist yet, initialize these values in the source hash with an
empty hash or array, respectively for `GDK::ConfigType::Settings` and
`GDK::ConfigType::SettingsArray`.

Because of this introduction of setting an empty hash or array, we also
need to change the definition of `GDK::ConfigSettings#user_defined?`.
This method now checks if any of the subsettings is set. This will
ensure that a subsettings set to `{}` will not make the parent setting
think a user-value is set.