Listen to fewer Sidekiq queues
This uses the new routing rules feature to route all workers (except a
few problem cases) to the `default` queue, instead of each job having
its own queue. It also sets the Sidekiq process to only listen to that
specific list of named queues.

The intention here is to make it less likely that someone adds a new
worker that requires its own queue - we're trying to clean up the
existing cases - so that we can use this configuration for self-managed
instances by default too.