Add explicit /dev/stderr to nginx cmdline
According to https://trac.nginx.org/nginx/ticket/147 the default error
log path compiled in the nginx binary is `/var/log/nginx/error.log` which
is opened before the configuration file is parsed.

If such a binary is used it will fail to start with Permission denied so
we have to pass /dev/stderr via -e option.