Allow GDK to manage Docker runners
This change allows GDK to manage a containerized Gitlab Runner.
Currently, the Docker configuration described in doc/howto/runner.md
requires that the user manually control the runner with docker run,
docker stop, docker restart, docker logs, etc. With this change, we can
have the runner container be managed alongside other GDK services with
gdk start and so on.