Document how to make GraphQL Explorer work
The GraphQL Explorer feature depends on the `graphiql-rails` gem, which
produces its own static JavaScript and CSS assets. These assets don't
appear to be generated on-demand for some reason, probably because we've
disconnected Sprockets. For now, we can document the workaround for
precompiling the assets and link the issue for removing this gem.