Fix gdk update creating an extra "gdk\n" directory
With $CDPATH not empty, cd prints the directory it moved to, if the path
was in $CDPATH. Since to use $CDPATH to keep it's default behaviour,
usually the users add .(current directory) as the top of the $CDPATH,
like: .:other/path/after.

Since gdk uses `cd $(dirname $0); pwd -P` to find the GDK_ROOT path, if
the user have value in $CDPATH, the output of the command could be
duplicated, generating an extra folder.

Related to: https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1400