Fix git diff spew in `gdk reconfigure`
`gdk reconfigure` attempts to show the diff between rendered ERB
files, but if the console doesn't detect colors, the `git diff` call
gets passed with an extra, blank argument that is rejected by the
process.

It's not clear the color detection works properly in macOS, but we can
at least fix the issue by passing `--no-color` if colors are not
available.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1418