Clarify shell and Docker executor runner setup
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/2386
dropped the setup steps for the Docker configuration, but it wasn't
clear the registration process was the same for both shell and Docker
executors.

We got rid of the use of registering with the Docker container because
it's easier to download the binary and have it write the initial
config file directly into the local path.