Build Gitaly binaries without GNU build ID
Gitaly has recently grown the ability to build its binaries with a GNU
build ID to be able to uniquely resolve debug symbols. This is slowing
down the build by quite a bit though given that Gitaly's commands aren't
compiled in parallel anymore but sequentially.

We don't need the build ID in the context of the GDK, and the slowdown
is quite visible when rebuilding Gitaly. So let's disable generation of
build IDs to speed it up again.