Drop libpqxx as a package dependency
libpqxx was added as a dependency for Fedora in d46963645, but as far
as I know GitLab or any of its dependencies do not depend on C++
bindings for PostgreSQL. Note that libpq, a C library for PostgreSQL,
is different from libpqxx.

Since this package isn't available in RHEL 8.3 with Extra Packages for
Enterprise Linux (EPEL), and it doesn't appear that Debian and Ubuntu
need this library as a dependency, let's drop it.