gitlab: Use Praefect in test when enabled
It's configurable to use Praefect in GDK. But that only applies to
development. In test Gitaly is always used without Praefect.

With this change tests are also using Praefect when the above setting is
enabled. In CI all tests already run with Praefect enabled [1] & [2], so
having it enabled locally can help detect possible issues earlier.

1. https://gitlab.com/gitlab-org/gitlab/-/blob/31e2dbe43f67d358107962726b413430d95bc465/config/gitlab.yml.example#L1527
2. https://gitlab.com/gitlab-org/gitlab/-/blob/31e2dbe43f67d358107962726b413430d95bc465/scripts/prepare_build.sh#L10

Changelog: changed