Refactor RailsMigration into TaskHelpers
This creates a designated area for Helper classes used in
tasks.

Having it separate from the main flow is important as those
helpers have different flows and expectations.

A task helper can terminate the execution with an `exit()`
code with makes it not suitable for use as part of the
normal GDK API.