gitaly: Migrate from internal socket dir to runtime dir
Gitaly has recently introduced a new runtime directory that is supposed
to hold all files and directories that are required at runtime. This was
introduced to fix issues with systems which automatically prune files in
`/tmp`, which is where Gitaly previously created some of its files.

This new setting superseded support for the `internal_socket_dir` config
which has been consequentially deprecated and removed for v15.0. In the
process, it was overseen that the GDK is still using the internal socket
directory though.

Migrate the GDK from using the internal socket directory to instead
configure the new runtime directory.