gitaly: Add diagnostics for internal socket path length
Gitaly is creating internal sockets in the runitme directory. This
runtime path directory can be nested quite deep, which overall may
result in a rather long socket path length. Unfortunately, many Unix
systems have length limitations in place for sockets: macOS has a limit
of 104 characters.

It is quite easy to hit this limit. We thus have a diagnostic in
Praefect which verifies that the internal socket directory doesn't
exceed a certain length limit. Funny enough though, Praefect doesn't
actually have an internal socket directory at all. So we have one
pointless check for Praefect, but no check for Gitaly.

Let's implement this diagnostic for Gitaly to alert users early on in
case they're about to exceed the limit. We'll remove support for the
internal runtime directory in Praefect in a subsequent commit.