Config: Stop reading from loose files
Configuring GDK through gdk.yml was introduced in the summer of 2019[1].
It's time we make this the only way to configure GDK and get rid of the
use of loose files.

With these changes the GDK config will raise an error if one of the old
loose files is detected. An error might look something like this:

    WARNING: WARNING: Loose file 'webpack_port' is no longer supported.
    INFO: Instead, set 'webpack.port' in your gdk.yml.

[1]: 4a85f1c4 (Add some ruby code to load config from yaml, 2019-03-18)

Changelog: deprecated