Simplify NGINX loopback interface instructions
The previous NGINX instructions may lead to confusion since
instructions for 127.1.1.1 don't discuss how to ensure that address
stays around after a reboot. We already have instructions in
`local_network.md` to use `127.16.123.1`, so let's standardize on that
address.

Also add a note that the step can be skipped unless you need to use
a runner under Docker.