Ensure `ruby --version` is run with a fresh PATH in pre-flight check
If a `gdk update` pulls in an update to `.tool-versions` that changes
the Ruby interpreter, the pre-flight dependency check may erroneously
report an outdated Ruby version due to a stale PATH set with the prior
version.

To avoid this error, we call `asdf exec ruby --version` so that the
PATH will be set to whatever version is specified in the updated
`.tool-versions`.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1618