Print Homebrew missing message to stderr
Previously when Homebrew were not installed, running the GDK install
might fail mysteriously due to:

```
support/makefiles/Makefile.test.mk:73: *** target pattern contains no `%'.  Stop.
```

If Homebrew is missing, previously the error message, `ERROR: Homebrew
is required but cannot be found.` would be sent to stdout, where it
would be consumed by Makefiles as valid input.

To avoid this, we send the output to stderr so that Makefiles will
not see any stdout leaking.