Fix support/component-git-clone checkout revision
This command was meant to checkout the specified revision but the
`git checkout` command was running in the context of the current working
directory. This would never work because the repository was just cloned
into a subdirectory. Checking all usages of this command it seems we
always provide the name of the directory being cloned into so we can
safely assume `ARGV.last` is the name of the directory we just cloned
it.

Prior to this change the logic in `def checkout` was doing nothing and
possibly places where this is being used also made use of other steps to
update the SHA which made this difficult to notice.