Add support for automatically installing and running Zoekt
The Zoekt integration was recently added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105049 . This
makes it easy to configure Zoekt in your GDK environment.

Unlike Postgres, Zoekt, unfortunately, cannot share a single server for
dev/test environment so we need to run separate servers for each. Then
Zoekt is also composed of an indexserver and webserver which are
separate so in total this means running 4 separate web servers in GDK.