Disable use of system libffi for building ffi gem
The ffi gem will attempt to use the system ffi library if the
development headers and libraries are available. Otherwise, the gem
will compile and install its own static library. For the GDK, the
latter approach is advantageous since the GDK won't break if the
system's version is upgraded.

We ran into a similar issue with Omnibus GitLab:
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/6800

Relates to:

1. https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/891
2. https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1783