dependencies: Use psql from postgresql.bin_dir
The user can configure `postgresql.bin_dir` in their `gdk.yml`. This
bypasses the scripts to auto-detect the location of `psql`. And in
52599d0a4 (support: Use pg_config from gdk config, 2023-06-08) we
modified the bootstrap scripts to properly use `psql` from
`postgresql.bin_dir`.

To check the dependencies, this `bin_dir` also should be used, instead
of searching for a possible location ourselves.

Changelog: fixed