Automatically rebuild pgvector if pg_config flags changed
When upgrading PostgreSQL, the pgvector library needs to be completely
cleaned and rebuilt for the extension to load properly.  This commit
runs `pg_config` and caches the output to `tmp/.pg_flags`.  If the
values changed, then `make clean` will be run inside the `pgvector`
directory.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3085#note_1391098623

Changelog: changed