Prevent `gdk tail` from hanging with unknown services
We had code to return warnings but the specs bypassed the method to
fetch the logs so we never actually triggered those warnings and we
ended up hanging as ruby was attempting to tail files that don't exist.

I've added an `.exist?` call to check the Pathname does actually exist
before attempting to tail it.