postgresql: Allow configuration of `max_connections`
By default, PostgreSQL allows up to a 100 connections at the same time,
and any connection made after this limit is hit is getting rejected. In
the context of the GDK a 100 connections should be plenty of headroom
for a normal installation.

However, the PostgreSQL server is also used by Gitaly's test suite. When
running tests with the Praefect proxy enabled, many of our tests will
spin up a separate Praefect process that then connects to the database.
Due to Gitaly's tests being highly parallelized nowadays, this easily
causes the tests to hit the connection limit.

Make the connection limit configurable such that Gitaly developers can
bump it as required. The default limit of a 100 connections is not
changed.

Changelog: added