Fix `make pgvector-setup` when pgvector/vector.so not present
If `pgvector/vector.so` is not present, users will see:

```
make: *** No rule to make target `pgvector/vector.so', needed by
`/Users/myuser/.asdf/installs/postgres/13.9/lib/vector.so'.  Stop.
```

That's because there is no rule for `pgvector/vector.so`, caused by a
regression introduced in https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3273.

To fix this, re-add the rule to `pgvector/vector.so` but just make
the installed library depend on that.