Skip truncation of database tables if CI database not enabled
If you have `gitlab.rails.databases.ci.enabled` set to `false`, then
the Rake tasks added in
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3272
will fail with an error since the main and CI database are the same.

This commit skips the truncation step if the CI database is not
enabled.

Relates to https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1905