Ensure PostgreSQL is stopped gracefully in build-verify-image
In the build-verify-image CI job, we do a lot of database work, such
as seeding the database and truncating tables, and then abruptly kill
all processes. This leads to many outstanding write-ahead log (WAL)
records, causing a long recovery process to occur after the database
is started up next.

This was discovered in
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3308#note_1507154071.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1910