gitaly: Migrate away from [gitlab-shell] configuration
Due to historic reasons, Gitaly still has a `[gitlab-shell]` section in
its configuration file. This section is used to configure two different
things:

  - The location of the GitLab secret used to connect to the GitLab
    server.

  - The location of custom hooks.

Gitaly has introduced replacements for both of these usecases quite a
while ago already so that it can drop the reference to `gitlab-shell`
eventually. And starting with 16.4, it doesn't require the old section
to exist anymore.

Migrate our Gitaly configuration template to use these replacements so
that Gitaly can deprecate and eventually remove the `[gitlab-shell]`
configuration. As Gitaly verifies the existence of the custom hooks
directory in case it has been explicitly configured we also make sure to
create that directory now.