Ensure pgvector library is installed in PostgreSQL (try 2)
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3273.
That merge request does not appear to be enough anymore.

This fixes the problem where moving the library out of the way doesn't
automatically reinstall:

```shell
mv `pg_config --libdir`/vector.so /tmp
make pgvector-setup
```

This should result in:

```
make: Nothing to be done for `pgvector-setup'.
```

On this branch, you should see:

```shell
% make pgvector-setup

--------------------------------------------------------------------------------
Installing pgvector v0.4.1 in /Users/stanhu/.asdf/installs/postgres/14.9/lib/vector.so
--------------------------------------------------------------------------------
```