command: Ensure databases are setup in `gdk reset-data`
In 58bf8ebc0 (Introduce a new target `ensure-databases-setup`,
2023-03-03) the prerequisites for the target `ensure-databases-running`
were moved to a new target `ensure-databases-setup`. This is great for
most cases, because in most situations the prerequisites are ready
anyway.

For the `gdk reset-data` command it requires to set up the databases
again, so swap out the target to `ensure-databases-setup`.

Label: bug::functional