services: Convert rails-web to a pure Ruby service
The service rails-web has a lot of environment variables, so it has
become tedious to maintain those. Convert this service to a pure Ruby
service, which allows setting environment variables as a hash.

Because pure Ruby services are still experimental, add a method so the
command can easily be rendered into the Procfile.