Fix building ffi gem on M1
In a previous commit [1], we switched to using the built-in FFI
provided with the `ffi` Ruby gem to avoid problems with upgrades
to the system FFI breaking the gem. However, this doesn't always
work because the built-in FFI implementation doesn't always build
correctly, dependent on the system compiler. For example, many
M1 Macs with a recent XCode installation are seeing odd failures
since 2022 [2].

In this commit, the switch to use `--disable-system-ffi` is reveted;
however, we still won't be using the system FFI. Instead, we will
take advantage of the fact that we use a `Brewfile` and have
Homebrew install a stable `libffi` version for us.

[1] https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/3061
[2] https://github.com/ffi/ffi/issues/937