Add libffi development packages to requirements
As described in https://bugs.ruby-lang.org/issues/18571, Ruby 3.2
stopped bundling libffi to build the `fiddle` gem. As a result,
`fiddle` is now an optional dependency if the `libffi` and its source
headers are not present.

This causes Ruby interpreters that are built without the
necessary dependencies to omit the `fiddle` C extension.

Make this a dependency in the GDK to ensure Ruby builds
properly.

Relates to:

* https://gitlab.com/gitlab-org/gitlab/-/issues/438901
* https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/2035

Changelog: fixed