Statically link libicu for macOS for static_holmes gem
static_holmes is a fork of the charlock_holmes gem with a single
change to make it possible to statically link icu4c libraries via the
`--enable-static` config option:
https://gitlab.com/gitlab-org/ruby/gems/charlock_holmes/-/merge_requests/2

Note that this only works on macOS and Homebrew. Ubuntu Linux doesn't
compile the static library with position-independent code (`-fPIC`),
so link errors result if `--enable-static` is used.

Relates to https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/2062